+++ resources/js/humhub.text_editor.js
@@ -3,6 +3,7 @@ humhub.module('text_editor', function (module, require, $) {
     var modal = require('ui.modal');
     var object = require('util').object;
     var Widget = require('ui.widget').Widget;
+    var event = require('event');
 
     var Editor = function (node, options) {
         Widget.call(this, node, options);
@@ -38,8 +39,28 @@ humhub.module('text_editor', function (module, require, $) {
         evt.finish();
     }
 
+    var createSubmit = function (evt) {
+        client.submit(evt).then(function (response) {
+            var modalWindow = modal.get('#texteditor-modal');
+            if (response.success) {
+                event.trigger('humhub:file:created', [response.file]);
+                if (response.editFormUrl) {
+                    modalWindow.load(response.editFormUrl);
+                    modalWindow.show();
+                } else {
+                    modalWindow.close();
+                }
+            } else {
+                modalWindow.setContent(response.output);
+            }
+        }).catch(function (e) {
+            module.log.error(e, true);
+        });
+    };
+
     module.export({
         Editor: Editor,
+        createSubmit: createSubmit,
     });
 
 });