+++ docs/CHANGELOG.md
@@ -4,5 +4,6 @@ Changelog
 0.0.1 - Unreleased
 -------------------------
 - Init: Initialize Module "Text editor"
-- Enh #1: Store file version on update content
 - Fix #5: Position of "Edit" Handler
+- Enh #2: Create new text files
+- Enh #1: Store file version on update content